+++ R/Learner_properties.R
@@ -45,6 +45,13 @@ getLearnerProperties.character = function(learner) {
   getLearnerProperties(checkLearner(learner))
 }
 
+#' @export
+getLearnerProperties.ModelMultiplexer = function(learner) {
+  selected = learner$par.vals$selected.learner
+  # NB: this is not set during construction
+  if (is.null(selected)) learner$properties else getLearnerProperties(learner$base.learners[[selected]])
+}
+
 #' @rdname LearnerProperties
 #' @export
 hasLearnerProperties = function(learner, props) {